split out node into a setup file
